Reorganize tests to match source code organisation
