+++ npm/src/cocooned/triggers.js
@@ -1,0 +1,7 @@
+import { Add } from './triggers/add'
+import { Remove } from './triggers/remove'
+
+export {
+  Add,
+  Remove
+}
